refactor(scripts): #72 2.2-2 — extract Probe/RowDiff/ProbeResult to scripts/validate/_shared.py
Move the three structurally cross-cutting dataclasses out of the
3,981-LOC validator into a 70-LOC shared module. Per-source planner +
runner blocks (TE, EODHD, BLS, BEA, Census, ISM, U Michigan, Conf
Board, NAR, ECB, Fed, NBS, METI, Stat Bureau JP) follow in slices
2.2-3..2.2-7.

Decoupling note: `Probe.expected_fields` previously defaulted to
`TE_PARSER_READS` (a TE-specific frozenset). The shared dataclass
now defaults to `frozenset()`, and the four TE planner probes that
relied on the implicit default get an explicit
`expected_fields=TE_PARSER_READS`. Field-diff behavior is identical
for TE; METI / Stat Bureau probes never read this field, so the
default flip is invisible there.

Adds `sys.path.insert(0, str(REPO_ROOT))` so the validator can do
`from scripts.validate._shared import ...` whether it is run as a
script or loaded by tests via `importlib.spec_from_file_location`.

Tests: 9 scaffold tests that load the validator unchanged at 135
passed; broader `-k "scaffold or validate or scripts"` subset at 695
passed (matches 2.2-1 baseline).

Refs #72.
